implements utils.get_data
