+++ .gitattributes
@@ -9,7 +9,7 @@
 *.ico binary
 *.pdf binary
 # Explicit LFS tracking for test files
-tests/data/.lfs/*.tar.gz filter=lfs diff=lfs merge=lfs -text
+/data/.lfs/*.tar.gz filter=lfs diff=lfs merge=lfs -text
 *.onnx filter=lfs diff=lfs merge=lfs -text binary
 *.mp4 filter=lfs diff=lfs merge=lfs -text binary
 *.mov filter=lfs diff=lfs merge=lfs -text binary
