+++ dimos/perception/segmentation/sam_2d_seg.py
@@ -12,25 +12,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import cv2
+import os
 import time
+from collections import deque
+from concurrent.futures import ThreadPoolExecutor
+
+import cv2
+import onnxruntime
 from ultralytics import FastSAM
+
+from dimos.perception.common.detection2d_tracker import get_tracked_results, target2dTracker
+from dimos.perception.segmentation.image_analyzer import ImageAnalyzer
 from dimos.perception.segmentation.utils import (
+    crop_images_from_bboxes,
     extract_masks_bboxes_probs_names,
     filter_segmentation_results,
     plot_results,
-    crop_images_from_bboxes,
 )
+from dimos.utils.data import get_data
 from dimos.utils.gpu_utils import is_cuda_available
-from dimos.perception.common.detection2d_tracker import target2dTracker, get_tracked_results
-from dimos.perception.segmentation.image_analyzer import ImageAnalyzer
-import os
-from collections import deque
-from concurrent.futures import ThreadPoolExecutor
 from dimos.utils.logging_config import setup_logger
 from dimos.utils.path_utils import get_project_root
-import onnxruntime
-from dimos.utils.testing import testData
 
 logger = setup_logger("dimos.perception.segmentation.sam_2d_seg")
 
@@ -55,7 +57,7 @@ class Sam2DSegmenter:
             logger.info("Using CPU for SAM 2d segmenter")
             self.device = "cpu"
         # Core components
-        self.model = FastSAM(testData(model_path) / model_name)
+        self.model = FastSAM(get_data(model_path) / model_name)
         self.use_tracker = use_tracker
         self.use_analyzer = use_analyzer
         self.use_rich_labeling = use_rich_labeling
